requirements.txt
amqp==2.5.2
anyjson==0.3.3
astroid==2.3.1
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.8.0
backcall==0.1.0
billiard==3.6.3.0
boto3==1.10.20
botocore==1.13.20
celery==4.4.2
certifi==2019.9.11
chardet==3.0.4
coverage==5.0.3
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.5
django-cors-headers==3.2.0
django-cprofile-middleware==1.0.5
django-crispy-forms==1.8.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-sslserver==0.21
django-storages==1.7.2
djangorestframework==3.11.0
docutils==0.15.2
factory-boy==2.12.0
Faker==4.0.0
flamegraph==0.1
flower==0.9.3
gprof2dot==2019.11.30
graphviz==0.13.2
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
jmespath==0.9.4
kombu==4.6.8
lazy-object-proxy==1.4.2
mccabe==0.6.1
mock==4.0.1
more-itertools==7.2.0
oauthlib==3.1.0
packaging==19.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pluggy==0.13.0
prompt-toolkit==2.0.10
psycopg2==2.8.3
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.4.2
pyparsing==2.4.2
pytest==5.1.3
python-dateutil==2.8.0
python-magic==0.4.15
python-memcached==1.59
python3-openid==3.1.0
pytz==2019.2
redis==3.4.1
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.2.1
selenium==3.141.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
text-unidecode==1.3
tornado==5.1.1
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.7
vine==1.3.0
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
